TransitPulse/requirements.txt
requests==2.31.0
confluent-kafka==2.3.0
python-dotenv==1.0.0
# apache-flink==1.17.1  # Incompatible with Python 3.13 locally; run in Docker
psycopg2-binary==2.9.9
redis==5.0.1
